+++ spec/helpers/application_helper_spec.rb
@@ -182,7 +182,7 @@ RSpec.describe ApplicationHelper do
       let(:service) { 'management_consultancy' }
 
       it 'returns the correct link with text' do
-        expect(result).to eq '<a target="blank" class="govuk-link" rel="noreferrer noopener" href="https://www.crowncommercial.gov.uk/contact?service=Management+Consultancy">Contact us</a>'
+        expect(result).to eq '<a target="blank" class="govuk-link" rel="noreferrer noopener" href="https://www.gca.gov.uk/contact?service=Management+Consultancy">Contact us</a>'
       end
     end
 
@@ -190,7 +190,7 @@ RSpec.describe ApplicationHelper do
       let(:service) { 'management_consultancy/admin' }
 
       it 'returns the correct link with text' do
-        expect(result).to eq '<a target="blank" class="govuk-link" rel="noreferrer noopener" href="https://www.crowncommercial.gov.uk/contact?service=Management+Consultancy">Contact us</a>'
+        expect(result).to eq '<a target="blank" class="govuk-link" rel="noreferrer noopener" href="https://www.gca.gov.uk/contact?service=Management+Consultancy">Contact us</a>'
       end
     end
 
@@ -198,7 +198,7 @@ RSpec.describe ApplicationHelper do
       let(:service) { 'legal_services' }
 
       it 'returns the correct link with text' do
-        expect(result).to eq '<a target="blank" class="govuk-link" rel="noreferrer noopener" href="https://www.crowncommercial.gov.uk/contact?service=Legal+Services">Contact us</a>'
+        expect(result).to eq '<a target="blank" class="govuk-link" rel="noreferrer noopener" href="https://www.gca.gov.uk/contact?service=Legal+Services">Contact us</a>'
       end
     end
 
@@ -206,7 +206,7 @@ RSpec.describe ApplicationHelper do
       let(:service) { 'legal_services/admin' }
 
       it 'returns the correct link with text' do
-        expect(result).to eq '<a target="blank" class="govuk-link" rel="noreferrer noopener" href="https://www.crowncommercial.gov.uk/contact?service=Legal+Services">Contact us</a>'
+        expect(result).to eq '<a target="blank" class="govuk-link" rel="noreferrer noopener" href="https://www.gca.gov.uk/contact?service=Legal+Services">Contact us</a>'
       end
     end
 
@@ -214,7 +214,7 @@ RSpec.describe ApplicationHelper do
       let(:service) { 'legal_panel_for_government' }
 
       it 'returns the correct link with text' do
-        expect(result).to eq '<a target="blank" class="govuk-link" rel="noreferrer noopener" href="https://www.crowncommercial.gov.uk/contact?service=Legal+Panel+for+Government">Contact us</a>'
+        expect(result).to eq '<a target="blank" class="govuk-link" rel="noreferrer noopener" href="https://www.gca.gov.uk/contact?service=Legal+Panel+for+Government">Contact us</a>'
       end
     end
 
@@ -222,7 +222,7 @@ RSpec.describe ApplicationHelper do
       let(:service) { 'legal_panel_for_government/admin' }
 
       it 'returns the correct link with text' do
-        expect(result).to eq '<a target="blank" class="govuk-link" rel="noreferrer noopener" href="https://www.crowncommercial.gov.uk/contact?service=Legal+Panel+for+Government">Contact us</a>'
+        expect(result).to eq '<a target="blank" class="govuk-link" rel="noreferrer noopener" href="https://www.gca.gov.uk/contact?service=Legal+Panel+for+Government">Contact us</a>'
       end
     end
 
@@ -230,7 +230,7 @@ RSpec.describe ApplicationHelper do
       let(:service) { 'supply_teachers' }
 
       it 'returns the correct link with text' do
-        expect(result).to eq '<a target="blank" class="govuk-link" rel="noreferrer noopener" href="https://www.crowncommercial.gov.uk/contact?service=Supply+Teachers">Contact us</a>'
+        expect(result).to eq '<a target="blank" class="govuk-link" rel="noreferrer noopener" href="https://www.gca.gov.uk/contact?service=Supply+Teachers">Contact us</a>'
       end
     end
 
@@ -238,7 +238,7 @@ RSpec.describe ApplicationHelper do
       let(:service) { 'supply_teachers/admin' }
 
       it 'returns the correct link with text' do
-        expect(result).to eq '<a target="blank" class="govuk-link" rel="noreferrer noopener" href="https://www.crowncommercial.gov.uk/contact?service=Supply+Teachers">Contact us</a>'
+        expect(result).to eq '<a target="blank" class="govuk-link" rel="noreferrer noopener" href="https://www.gca.gov.uk/contact?service=Supply+Teachers">Contact us</a>'
       end
     end
 
@@ -246,7 +246,7 @@ RSpec.describe ApplicationHelper do
       let(:service) { 'auth' }
 
       it 'returns the correct link with text' do
-        expect(result).to eq '<a target="blank" class="govuk-link" rel="noreferrer noopener" href="https://www.crowncommercial.gov.uk/contact?service=Supply+Teachers">Contact us</a>'
+        expect(result).to eq '<a target="blank" class="govuk-link" rel="noreferrer noopener" href="https://www.gca.gov.uk/contact?service=Supply+Teachers">Contact us</a>'
       end
     end
   end